Disobedience no longer attempts to compute defaults for namepart preferences.  It doesn't in general have enough information to do so correctly.
For fetching the preferences, with defaults still working, it now just
asks for the namepart and infers that the preference matches it.

For setting the preferences it no longer attempts to unset preferences
that are now at their default value.

This will cause trivial preferences to accumulate in the database.
This can be fixed if desired by making the server translate such sets
into unsets.