In the web interface only, new 'short' track name part context.  This is the same as 'display' but truncated to short_display glyphs.
playing, recent and new now use this to bound the width of columns in
their tables.  Hover over the truncated names to see the full name (in
a TITLE attribute).

As part of this, rewrote the unicode support.  Everything is now done
via unidata.h/c, which are generated by the new scripts/make-unidata
from UnicodeData.txt (which is not included; get the latest version
from unicode.org instead if you need to update these files).  The
result is both simpler and probably also faster, although it uses more
(immutable) memory.