Start of Unicode support rewrite
unicode.c/.h provide a new (not yet complete) Unicode API which implements
algorithms as described in the Unicode 5.0.0 standard.  lib/test and words.c
are somewhat modified to use them.

By default the make-unidata script is NEVER run.  It must be run manually.
This policy might be changed in the future; the script is capable of
downloading any files it needs to operate.