unicode.c iterators can now have their notion of Word_Break tailored.
We use this to arrange that underscores are treated as spaces, since a
common track naming convention replaces spaces with underscores.  We
only apply this to raw filenames, not to the overrides provided by
preferences, on the assumption that if you typed an underscore there
you really meant it.

words.c/h are now gone; all references to words() and casefold() use
unicode.c functions directly.