disorder-rescan can now suppress the check phase, which on first startup with lots of tracks is rather time consuming.
Additionally, it avoids doing most of the work inside a transaction,
by the rather brute expedient of pulling the entire track list into
memory and then iterating over that.

The server takes advantage of this by making the initial blocking
rescan not do the check phase, so it's reasonably quick, and then
issuing a second non-blocking rescan immediately.

It might be better all round to calculate track lengths on demand but
this arrangement should be better than previously.