Sort out static content.  There is a new expansion @image:NAME@ which produces the URL for the static content called NAME (or NAME.png if you don't supply an extexnsion).  The labels images.NAME can be used to indirect role names.
url.static still has its old meaning but does not appear in
options.labels as it has a sensible default.

This is then used throughout to find images and the stylesheet.

The main options file is now the same for Debian and non-Debian.  On
Debian a deafult empty options.user file is installed.

Updated a lot of copyright dates.