disorder authorize now creates ~USER/.disorder/passwd instead of /etc/disorder/config.USER.  This is a bit fiddlier than the old code but reduces the number of places the user must look for their password.
Generated passwords are base64 rather than hex, and are slightly
longer (in terms of the underlying bytes).  They aren't human
memorable but in practice they are expected to be kept in
~/.disorder/passwd or a browser's password store.

The configuration parser now always uses pw_home rather than $HOME to
find ~/.disorder.

Mention Ubuntu in README as the debs work there too.