Make the Disobedience login window a bit saner.  There is now just a Login and a Cancel button.  The former attempts a login synchronously (there not being any background activity to worry about blocking at this point); on success it (unconditionally) saves the config and pops down the login window, on error it gives you another go.
The problem described in defect 17 is eliminated in the login box by
this change too since polling of the eclient is suppressed while the
login box is up.  However the underlying problem is still there and
probably relates to multiple commands failing due to authentication
errors.  Those commands should probably be discarded silently.