uaudio gains a new 'configure' method, which imposes the audio API's configuration from the global (configuration.h) config.  This gets network play working again.
There's a new configuration option, rtp_delay_threshold, which allows
the courageous to influence the internal timings of the RTP
transmitter.  The defaults seem to work however.

The reason it doesn't go directly to global config is that uaudio
needs to be usable from playrtp, which might want to be set to totally
different settings to whatever the local DisOrder server instance
happens to be.