Rewrite uaudio-schedule.c.  It's now much simpler and should actually get the timing right.  An hour of rtpmon is encouraging.
This change pushes responsibility for pausing out towards the APIs.

For ALSA we sleep a bit.  For OSS we close the sound device and sleep a 
bit.

For RTP we just pretend to send packets; the scheduler takes care of the 
timing.

For command again the scheduler takes care of the timing, though as 
before we have the option of either suspending writes or sending 0s.

Scantily tested!