playrtp: refuse to play RTP via RTP.
playrtp attempting to use the RTP backend becomes the default
situation if no local sound output API is available.  This can happen
using (e.g.)  Fink's GCC 4.4, because it cannot compile Apple's header
files, and so concludes that CoreAudio/AudioHardware.h is not
available.

This is about the minimal possible change to stop playrtp crashing in
this case.  There are a variety of better possible answers.