server/gstdecode.c: New program, like `disorder-decode'.
This uses GStreamer for the decoding, so it should be able to decode
more or less anything you can throw at it, assuming you have the
necessary plugins lying about.

More importantly (from my point of view) it does ReplayGain (optionally,
but by default).  This uses the per-album setting by default, but this
can be overridden.