tests/dtest.py: Maintain a separate test root for each test.
Now they can run in parallel without breaking each other, which is nice.
It's nice for two reasons: firstly, Automake changed behaviour and runs
tests in parallel by default nowadays; and secondly, the tests are
rather slow because they involve a bunch of waiting around for things,
and waiting for stuff in parallel is easy.

The downside is that we end up gobbling a lot of disk space with all of
the logfiles.  I think this is a good trade.