plugins/tracklength-gstreamer.c: Rewrite to use `GstDiscoverer'.
The old version worked, but was fiddly, and very slow at some MP3
files.  It seems that if you start playing for a few milliseconds and
then ask, then you get a good answer, but since I found this out by
tripping over the `GstDiscoverer' machinery that GStreamer already has,
it seemed like a better idea to just use that.  Which means almost all
of the code has gone.