clients/disorder.c: Force line buffering when writing log.
Otherwise it looks fine on a terminal, but that's not very useful.  If a
program tries to read the log over a pipe, then it sees nothing for a
good long while until the buffer gets flushed.

This change is specific to the `log' command.  Output for other commands
comes out at the same time anyway.