server/Makefile.am: Allow discrepancies for lossy codecs.
It seems that `libMAD' on ARM32 doesn't produce the same output as it
does on x86; and the ARM64 version is a looooong way off.  I've not
investigated, but I suspect details of floating point and/or SIMD
optimization.  Anyway, smallish $L_\infty$ distance isn't especially
concerning.  After all, we're trying to make sure that we're using the
decoders right, rather than that the decoders conform to the relevant
specifications.