clients/playrtp.c: Fix the RTP-stream request syntax.
Previously, the `-' syntax was an undocumented (and booby-trapped) way
to inform the rest of the program that it should request a unicast
stream.  If we see this, then we apply some heuristics to select an
interface to bind to, and ask the server to hurl packets at it.

This is unsatisfactory.  One might want to set a particular port, say
because other ports are firewalled off by default.  And one might want
to set a particular address if the heuristics go wrong.

As a first step, overhaul the command-line parser to accept an address
and port name/number.  We still apply the old heuristics if no address
is provided explicitly, and we get the kernel to select a port number if
none is given.

This change is rather larger than I usually like, but there doesn't seem
to be a useful intermediate state.  Sorry.