clients/playrtp.c, lib/configuration.[ch]: Static config for `playrtp'.
The `disorder-playrtp' program already reads the client configuration.
This change sets defaults for a number of parameters which could
previously only be changed from the command line, which is a problem
because `disorder-playrtp' is commonly invoked by Disobedience which
doesn't allow one to set a command-line.

These various tweaks are experimental, and may change, or even be
removed, in some later version.  The `rtp_request_address' and
`rtp_always_request' options, in particular, might be replaced by some
more complicated policy mechanism.