lib/configuration.c, lib/uaudio-rtp.c: Allow tweaking MTU-discovery.
Apparently under some circumstances, Linux tries to do path MTU
discovery with RTP transmissions, so it (a) sets DF on outgoing packets
and then (b) ignores the `fragmentation-needed' errors coming back!
This option attempts to work around this foolishness.

Configuration is accepted on all platforms for portability's sake, but
is only effective on Linux.